+++ src/app/profile/change-password/page.tsx
@@ -1,21 +1,25 @@
 import type { Metadata } from 'next';
-import { getProfileAction } from '@/app/profile/actions/get-profile-action';
+import { redirect, notFound } from 'next/navigation';
+import { getProfile } from '@/app/profile/data/get-profile';
 import { ChangePassword } from '@/app/profile/change-password/change-password';
-import {
-    isServerActionSuccess,
-    ServerActionError,
-} from '@/components/server-action-error';
+import { getSession } from '@/lib/session';
 
 export const metadata: Metadata = {
     title: 'Change Password',
 };
 
 export default async function ChangePasswordPage() {
-    const result = await getProfileAction();
+    const session = await getSession();
 
-    if (!isServerActionSuccess(result)) {
-        return <ServerActionError result={result} />;
+    if (!session) {
+        redirect('/account/login');
     }
 
-    return <ChangePassword profile={result.data} />;
+    const profile = await getProfile(session.sub);
+
+    if (!profile || profile.isLockedOut) {
+        notFound();
+    }
+
+    return <ChangePassword profile={profile} />;
 }
