+++ src/app/profile/update-profile/page.tsx
@@ -1,21 +1,25 @@
 import type { Metadata } from 'next';
-import { getProfileAction } from '@/app/profile/actions/get-profile-action';
+import { notFound, redirect } from 'next/navigation';
+import { getProfile } from '@/app/profile/data/get-profile';
 import { UpdateProfile } from '@/app/profile/update-profile/update-profile';
-import {
-    isServerActionSuccess,
-    ServerActionError,
-} from '@/components/server-action-error';
+import { getSession } from '@/lib/session';
 
 export const metadata: Metadata = {
     title: 'Update Profile',
 };
 
 export default async function UpdateProfilePage() {
-    const result = await getProfileAction();
+    const session = await getSession();
 
-    if (!isServerActionSuccess(result)) {
-        return <ServerActionError result={result} />;
+    if (!session) {
+        return redirect('/account/login');
     }
 
-    return <UpdateProfile profile={result.data} />;
+    const profile = await getProfile(session.sub);
+
+    if (!profile || profile.isLockedOut) {
+        return notFound();
+    }
+
+    return <UpdateProfile profile={profile} />;
 }
