+++ src/components/server-action-error.tsx
@@ -1,6 +1,4 @@
 import type { SafeActionResult } from 'next-safe-action';
-import { AlertCircleIcon } from 'lucide-react';
-import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert';
 
 type FlattenedSafeActionResult<T> = SafeActionResult<
     string,
@@ -9,24 +7,6 @@ type FlattenedSafeActionResult<T> = SafeActionResult<
     T,
     object
 >;
-
-export function ServerActionError<T>({
-    result,
-}: {
-    result?: FlattenedSafeActionResult<T>;
-}) {
-    return (
-        <div className="mx-auto max-w-screen-sm p-6">
-            <Alert variant="destructive">
-                <AlertCircleIcon />
-                <AlertTitle>Error</AlertTitle>
-                <AlertDescription>
-                    <ServerActionErrorMessage result={result} />
-                </AlertDescription>
-            </Alert>
-        </div>
-    );
-}
 
 export function ServerActionErrorMessage<T>({
     result,
@@ -52,22 +32,4 @@ export function ServerActionErrorMessage<T>({
         result?.serverError ||
         'An error occurred while processing your request.'
     );
-}
-
-export function isServerActionSuccess<T>(
-    result?: FlattenedSafeActionResult<T>
-): result is { data: T } {
-    if (!result) {
-        return false;
-    }
-
-    if (result.validationErrors) {
-        return false;
-    }
-
-    if (result.serverError) {
-        return false;
-    }
-
-    return true;
 }
